requirements.txt
ultralytics==8.3.20
opencv-python==4.10.0.84
supervision==0.24.0
Flask==3.0.3
numpy==1.26.4
pillow==10.3.0
streamlit==1.32.0
requests==2.32.2